refactor: pull assertVictory(), assertLoss() and assertStalemate functions into testing helpers file
